+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+* New cli and yaml (config) options `--exclude-paths DIR`.
+* Encrypted disk is no more default, use `--disk-encrypt` if need.
+
 ## 0.2.0, release 2023/10/23
 * Define constant OPTIONS and ACTIONS
 * Correct rsync options